Improve code
